Revert "Use int instead of long-int in fib() to resolve test errors in CI due to c-stack usage limit"
This reverts commit 25fd829e9a9912365b43500b910cfd93fc2085c4.

+++ src/RcppExports.cpp
@@ -12,23 +12,23 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
 // fibonacci
-std::vector<int> fibonacci(std::vector<int> n);
+std::vector<long int> fibonacci(std::vector<long int> n);
 RcppExport SEXP _rcpptimer_fibonacci(SEXP nSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< std::vector<int> >::type n(nSEXP);
+    Rcpp::traits::input_parameter< std::vector<long int> >::type n(nSEXP);
     rcpp_result_gen = Rcpp::wrap(fibonacci(n));
     return rcpp_result_gen;
 END_RCPP
 }
 // fibonacci_omp
-std::vector<int> fibonacci_omp(std::vector<int> n);
+std::vector<long int> fibonacci_omp(std::vector<long int> n);
 RcppExport SEXP _rcpptimer_fibonacci_omp(SEXP nSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< std::vector<int> >::type n(nSEXP);
+    Rcpp::traits::input_parameter< std::vector<long int> >::type n(nSEXP);
     rcpp_result_gen = Rcpp::wrap(fibonacci_omp(n));
     return rcpp_result_gen;
 END_RCPP
